notiservice/requirements.txt
azure-core==1.29.5
azure-servicebus==7.11.3
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pika==1.3.2
psycopg2-binary==2.9.9
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0
